99-archive/patient-experience-review/requirements.txt
requests
pytest
python-dotenv
llama-index
tiktoken 
chromadb
beautifulsoup4
markdown
deeplake
openai
langchain-core 
langchain-community 
langchain-openai
langchain-cli
langchainhub
pandas
datasets
ipywidgets